refactor: move course description component to separate file
